\name{valley-package}
\alias{valley-package}
\alias{valley}
\docType{package}
\title{
Various function for signal processing originaly tested on mass spectrometry data
}
\description{
Various function for signal processing originaly tested on mass spectrometry data.
}
\details{
\tabular{ll}{
Package: \tab valley\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-05-03\cr
License: \tab GPL-3 \cr
}
}
\author{
Julien Prados

Maintainer: Julien Prados <julien.pradost@unige.ch>
}
\references{
On Preprocessing of SELDI-MS Data and its Evaluation. Julien Prados, Alexandros Kalousis, Melanie Hilario. Computer-Based Medical Systems, 2006.
}
\keyword{ package }
\seealso{
\code{\link{valleys}} ~~
}
